Rename data structures and modules in farmer
