+++ crates/subspace-farmer/src/lib.rs
@@ -26,13 +26,13 @@ pub(crate) mod commitments;
 pub(crate) mod farming;
 pub(crate) mod identity;
 #[cfg(test)]
-mod mock_rpc;
+mod mock_rpc_client;
 pub mod multi_farming;
+pub(crate) mod node_rpc_client;
 pub(crate) mod object_mappings;
 pub(crate) mod plot;
 pub(crate) mod plotting;
-pub(crate) mod rpc;
-pub(crate) mod ws_rpc;
+pub(crate) mod rpc_client;
 pub mod ws_rpc_server;
 
 pub use archiving::{Archiving, ArchivingError, PiecesToPlot};
@@ -40,8 +40,8 @@ pub use commitments::{CommitmentError, Commitments};
 pub use farming::{Farming, FarmingError};
 pub use identity::Identity;
 pub use jsonrpsee;
+pub use node_rpc_client::NodeRpcClient;
 pub use object_mappings::{ObjectMappingError, ObjectMappings};
 pub use plot::{retrieve_piece_from_plots, Plot, PlotError};
 pub use plotting::plot_pieces;
-pub use rpc::RpcClient;
-pub use ws_rpc::WsRpc;
+pub use rpc_client::RpcClient;
